a SQL script that creates a stored procedure ComputeAverageScoreForUser that computes and store the average score for a student.

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -5,7 +5,7 @@ CREATE PROCEDURE ComputeAverageScoreForUser(
 	IN user_id INTEGER
 )
 BEGIN
-	DECLARE total_score DECIMAL(10,2);
+	DECLARE total_score INTEGER;
 	DECLARE total_corrections INTEGER;
 
 	SELECT SUM(score) INTO total_score
